Refactor site architecture and routing foundation
