+++ src/index.css
@@ -30,3 +30,17 @@ header.sticky > div > a[href="/"]:first-child::before {
   content: "";
   background: url('/brand/v1/ckconflux-symbol.svg') center / contain no-repeat;
 }
+
+:focus-visible {
+  outline: 2px solid #67e8f9;
+  outline-offset: 3px;
+}
+
+@media (prefers-reduced-motion: reduce) {
+  *, *::before, *::after {
+    scroll-behavior: auto !important;
+    animation-duration: 0.01ms !important;
+    animation-iteration-count: 1 !important;
+    transition-duration: 0.01ms !important;
+  }
+}
